test: migrate to mini.test
