+++ Makefile
@@ -1,4 +1,4 @@
-TESTS_INIT=tests/minimal.lua
+TESTS_INIT=scripts/minimal_init.lua
 TESTS_DIR=tests/
 DTS_SCRIPT=tests/precognition/dts.lua
 SEED_START=0
@@ -11,7 +11,7 @@ test:
 		--headless \
 		--noplugin \
 		-u ${TESTS_INIT} \
-		-c "PlenaryBustedDirectory ${TESTS_DIR} { minimal_init = '${TESTS_INIT}' }" \
+		-c "lua MiniTest.run()" \
 
 dts:
 	@nvim \
