+++ tests/precognition/test_gutter_hints.lua
@@ -1,8 +1,11 @@
 local precognition = require("precognition")
----@diagnostic disable-next-line: undefined-field
-local eq = assert.are.same
+local eq = MiniTest.expect.equality
 
 describe("Gutter hints table", function()
+    before_each(function()
+        precognition.setup({})
+    end)
+
     it("should return a table with the correct keys", function()
         local testBuf = vim.api.nvim_create_buf(true, true)
         vim.api.nvim_buf_set_lines(testBuf, 0, -1, false, {
