+++ tests/precognition/utils/compat.lua
@@ -1,5 +1,5 @@
 local M = {}
 
-M.get_active_lsp_clients = vim.lsp.get_clients() or vim.lsp.get_active_clients()
+M.get_active_lsp_clients = vim.lsp.get_clients or vim.lsp.get_active_clients
 
 return M
