+++ tests/screenshots/tests-precognition-test_e2e.lua---Gutter-Priority---0-priority-item-is-not-added
@@ -1,0 +1,51 @@
+--|---------|---------|---------|---------|---------|---------|---------|---------|
+01|ggABC                                                                           
+02|  DEF                                                                           
+03|{                                                                               
+04|  GHI                                                                           
+05|  ^ e                                                                           
+06|}                                                                               
+07|  JKL                                                                           
+08|                                                                                
+09|  MNO                                                                           
+10|~                                                                               
+11|~                                                                               
+12|~                                                                               
+13|~                                                                               
+14|~                                                                               
+15|~                                                                               
+16|~                                                                               
+17|~                                                                               
+18|~                                                                               
+19|~                                                                               
+20|~                                                                               
+21|~                                                                               
+22|~                                                                               
+23|[No Name] [+]                                                 4,1            All
+24|                                                                                
+
+--|---------|---------|---------|---------|---------|---------|---------|---------|
+01|00111111111111111111111111111111111111111111111111111111111111111111111111111111
+02|22111111111111111111111111111111111111111111111111111111111111111111111111111111
+03|00111111111111111111111111111111111111111111111111111111111111111111111111111111
+04|22111111111111111111111111111111111111111111111111111111111111111111111111111111
+05|22000111111111111111111111111111111111111111111111111111111111111111111111111111
+06|00111111111111111111111111111111111111111111111111111111111111111111111111111111
+07|22111111111111111111111111111111111111111111111111111111111111111111111111111111
+08|22111111111111111111111111111111111111111111111111111111111111111111111111111111
+09|22111111111111111111111111111111111111111111111111111111111111111111111111111111
+10|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+11|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+12|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+13|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+14|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+15|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+16|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+17|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+18|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+19|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+20|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+21|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+22|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+23|33333333333333333333333333333333333333333333333333333333333333333333333333333333
+24|44444444444444444444444444444444444444444444444444444444444444444444444444444444
