+++ tests/screenshots/tests-precognition-test_e2e.lua---e2e-tests---pads-full-width-virtual-lines-to-the-text-area
@@ -1,0 +1,51 @@
+--|---------|---------|---------|---------|---------|---------|---------|---------|
+01|G   1 Hello World                                                               
+02|      ^   e w   $                                                               
+03|~                                                                               
+04|~                                                                               
+05|~                                                                               
+06|~                                                                               
+07|~                                                                               
+08|~                                                                               
+09|~                                                                               
+10|~                                                                               
+11|~                                                                               
+12|~                                                                               
+13|~                                                                               
+14|~                                                                               
+15|~                                                                               
+16|~                                                                               
+17|~                                                                               
+18|~                                                                               
+19|~                                                                               
+20|~                                                                               
+21|~                                                                               
+22|~                                                                               
+23|[No Name] [+]                                                 1,1            All
+24|                                                                                
+
+--|---------|---------|---------|---------|---------|---------|---------|---------|
+01|00111122222222222222222222222222222222222222222222222222222222222222222222222222
+02|11111100000000000000000000000000000000000000000000000000000000000000000000000000
+03|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+04|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+05|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+06|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+07|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+08|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+09|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+10|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+11|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+12|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+13|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+14|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+15|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+16|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+17|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+18|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+19|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+20|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+21|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+22|11111111111111111111111111111111111111111111111111111111111111111111111111111111
+23|33333333333333333333333333333333333333333333333333333333333333333333333333333333
+24|44444444444444444444444444444444444444444444444444444444444444444444444444444444
