+++ tests/screenshots/tests-precognition-test_e2e.lua---e2e-tests---virtual-line-text-color-can-be-customised
@@ -1,0 +1,51 @@
+--|---------|---------|---------|---------|---------|---------|---------|---------|
+01|ggHello World this is a test                                                    
+02|  b   e w                  $                                                    
+03|  line 2                                                                        
+04|}                                                                               
+05|  line 4                                                                        
+06|                                                                                
+07|G line 6                                                                        
+08|~                                                                               
+09|~                                                                               
+10|~                                                                               
+11|~                                                                               
+12|~                                                                               
+13|~                                                                               
+14|~                                                                               
+15|~                                                                               
+16|~                                                                               
+17|~                                                                               
+18|~                                                                               
+19|~                                                                               
+20|~                                                                               
+21|~                                                                               
+22|~                                                                               
+23|[No Name] [+]                                                 1,2            All
+24|                                                                                
+
+--|---------|---------|---------|---------|---------|---------|---------|---------|
+01|00111111111111111111111111111111111111111111111111111111111111111111111111111111
+02|22000000000000000000000000001111111111111111111111111111111111111111111111111111
+03|22111111111111111111111111111111111111111111111111111111111111111111111111111111
+04|00111111111111111111111111111111111111111111111111111111111111111111111111111111
+05|22111111111111111111111111111111111111111111111111111111111111111111111111111111
+06|22111111111111111111111111111111111111111111111111111111111111111111111111111111
+07|00111111111111111111111111111111111111111111111111111111111111111111111111111111
+08|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+09|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+10|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+11|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+12|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+13|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+14|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+15|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+16|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+17|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+18|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+19|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+20|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+21|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+22|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+23|33333333333333333333333333333333333333333333333333333333333333333333333333333333
+24|44444444444444444444444444444444444444444444444444444444444444444444444444444444
