+++ tests/screenshots/tests-precognition-test_inlay_hints.lua---lsp-based-tests---inlay-hint-shifts-the-line
@@ -1,0 +1,51 @@
+--|---------|---------|---------|---------|---------|---------|---------|---------|
+01|G her fooe is a string                                                          
+02|  b      e w         $                                                          
+03|~                                                                               
+04|~                                                                               
+05|~                                                                               
+06|~                                                                               
+07|~                                                                               
+08|~                                                                               
+09|~                                                                               
+10|~                                                                               
+11|~                                                                               
+12|~                                                                               
+13|~                                                                               
+14|~                                                                               
+15|~                                                                               
+16|~                                                                               
+17|~                                                                               
+18|~                                                                               
+19|~                                                                               
+20|~                                                                               
+21|~                                                                               
+22|~                                                                               
+23|[No Name] [+]                                                 1,2            All
+24|vim.lsp.start_client() is deprecated. Run ":checkhealth vim.deprecat            
+
+--|---------|---------|---------|---------|---------|---------|---------|---------|
+01|00111122211111111111111111111111111111111111111111111111111111111111111111111111
+02|22000000000000000000001111111111111111111111111111111111111111111111111111111111
+03|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+04|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+05|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+06|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+07|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+08|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+09|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+10|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+11|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+12|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+13|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+14|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+15|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+16|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+17|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+18|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+19|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+20|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+21|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+22|22222222222222222222222222222222222222222222222222222222222222222222222222222222
+23|33333333333333333333333333333333333333333333333333333333333333333333333333333333
+24|44444444444444444444444444444444444444444444444444444444444444444444555555555555
